+++ src/lib/ggpo/network/udp_msg.h
@@ -27,8 +27,8 @@ struct UdpMsg
    };
 
    struct connect_status {
-      int         disconnected:1;
-      int         last_frame:31;
+      unsigned int   disconnected:1;
+      int            last_frame:31;
    };
 
    struct {
@@ -99,7 +99,7 @@ public:
       return 0;
    }
 
-   UdpMsg(MsgType t) { hdr.type = t; }
+   UdpMsg(MsgType t) { hdr.type = (uint8)t; }
 };
 
 #pragma pack(pop)
